refactor(nimbus): isolated new rollout design forms, views, and templates
