+++ experimenter/experimenter/nimbus_ui/templates/nimbus_experiments/new_designs/rollouts/detail_card.html
