--- experimenter/experimenter/nimbus_ui/templates/nimbus_experiments/new_rollout_detail.html
@@ -1,43 +1,0 @@
-{% extends "common/new_base.html" %}
-
-{% load static %}
-{% load nimbus_extras %}
-
-{% block title %}{{ experiment.name }}{% endblock %}
-{% block main_content %}
-  <div id="RolloutSummary" class="d-flex flex-column gap-4">
-    {# ── Define ── #}
-    <div class="d-flex flex-column gap-2">
-      <div class="small fw-medium mb-1 text-secondary">Define</div>
-      {# EXP-6681: Rollout Overview Form/Card #}
-      {% include "nimbus_experiments/rollout_cards/new_detail_card.html" with card_id="overview" title="Overview" subtitle="Name · Observations & Problem space · Important Links · Project Tags · Subscribers" show_edit_btn=True body_template="nimbus_experiments/overview/card.html" %}
-
-      {# TODO EXP-6682: Rollout Risks Form/Card #}
-      {% include "nimbus_experiments/rollout_cards/new_detail_card.html" with card_id="risks" title="Risks" subtitle="Consider possible risks if the public interacts with your experiment" show_edit_btn=True body_template="nimbus_experiments/risks/card.html" %}
-
-    </div>
-    {# ── Configure ── #}
-    <div class="d-flex flex-column gap-2">
-      <div class="small fw-medium mb-1 text-secondary">Configure</div>
-      {# TODO EXP-6683: Rollout Audience Form/Card #}
-      {% include "nimbus_experiments/rollout_cards/new_detail_card.html" with card_id="audience" title="Audience" subtitle="Branch treatment details" show_edit_btn=True body_template="nimbus_experiments/audience/card.html" %}
-
-    </div>
-    {# ── Build ── #}
-    <div class="d-flex flex-column gap-2">
-      <div class="small fw-medium mb-1 text-secondary">Build</div>
-      {# TODO EXP-6685: Rollout Experience Form/Card #}
-      {% include "nimbus_experiments/rollout_cards/new_detail_card.html" with card_id="features" title="Rollout experience" subtitle="Describe the rollout experience" %}
-
-      {# TODO EXP-6685: Rollout Schedule Form/Card #}
-      {% include "nimbus_experiments/rollout_cards/new_detail_card.html" with card_id="schedule" title="Rollout schedule" subtitle="1% → 10% → 50% → 100%" %}
-
-    </div>
-    {# ── Check ── #}
-    <div class="d-flex flex-column gap-2">
-      <div class="small fw-medium mb-1 text-secondary">Check</div>
-      {% include "nimbus_experiments/rollout_cards/new_detail_card.html" with card_id="qa" title="Quality Assurance Checks (QA)" subtitle="Request a QA or do a Self QA" show_edit_btn=True body_template="nimbus_experiments/qa/card.html" %}
-
-    </div>
-  </div>
-{% endblock %}
